% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{blastWA}
\alias{blastWA}
\title{Blast WhatsApp Messages to Multiple Number}
\usage{
blastWA()
}
\description{
Blast WhatsApp messages to multiple Number using template from Google Sheets.
}
\examples{
\dontrun{
# Make sure your Fonnte token is valid and Google Sheet has sheets named "Sheet1" and "Template"
# blastWA()
}
}
